moves projects into their new names, + updated tests projects (at each api level, onlyhave integ + unit)
