--- src/Mjolnir/ActionResults/UnworthyResult.cs
@@ -1,12 +1,0 @@
-using Microsoft.AspNetCore.Http;
-
-namespace Mjolnir.ActionResults
-{
-    public class UnworthyResult : CustomReasonResult
-    {
-        public UnworthyResult()
-        : base(StatusCodes.Status403Forbidden, "Unworthy")
-        {
-        }
-    }
-}
