--- src/Mjolnir/Configurations/BifrostConfiguration.cs
@@ -1,8 +1,0 @@
-namespace Mjolnir.Configurations
-{
-    public class BifrostConfiguration
-    {
-        public const string Key = "Bifrost";
-        public string Secret { get; set; }
-    }
-}
