--- src/Mjolnir/Infrastructure/AsgardianClaims.cs
@@ -1,7 +1,0 @@
-namespace Mjolnir.Infrastructure
-{
-    public static class AsgardianClaims
-    {
-        public static string Worthiness => "https://asgard.com/worthiness";
-    }
-}
