--- src/Mjolnir/Services/ICurrentHeroService.cs
@@ -1,7 +1,0 @@
-namespace Mjolnir.Services
-{
-    public interface ICurrentHeroService
-    {
-        bool WieldMjolnir();
-    }
-}
