--- src/Multiverse/Options/JwtOptions.cs
@@ -1,9 +1,0 @@
-namespace Multiverse.Options
-{
-    public class JwtOptions
-    {
-        public const string Key = "Jwt";
-
-        public string Secret { get; set; }
-    }
-}
