+++ src/Services/Asgard/ActionResults/UnworthyResult.cs
@@ -1,0 +1,12 @@
+using Microsoft.AspNetCore.Http;
+
+namespace Asgard.ActionResults
+{
+    public class UnworthyResult : CustomReasonResult
+    {
+        public UnworthyResult()
+        : base(StatusCodes.Status403Forbidden, "Unworthy")
+        {
+        }
+    }
+}
