+++ src/Services/Asgard/ActionResults/WorthyResult.cs
@@ -1,0 +1,12 @@
+using Microsoft.AspNetCore.Http;
+
+namespace Asgard.ActionResults
+{
+    public class WorthyResult : CustomReasonResult
+    {
+        public WorthyResult()
+        : base(StatusCodes.Status200OK, "Worthy")
+        {
+        }
+    }
+}
