+++ src/Services/Asgard/Configurations/BifrostConfiguration.cs
@@ -1,0 +1,8 @@
+namespace Asgard.Configurations
+{
+    public class BifrostConfiguration
+    {
+        public const string Key = "Bifrost";
+        public string Secret { get; set; }
+    }
+}
