+++ src/Services/Asgard/Infrastructure/AsgardianClaims.cs
@@ -1,0 +1,7 @@
+namespace Asgard.Infrastructure
+{
+    public static class AsgardianClaims
+    {
+        public static string Worthiness => "https://asgard.com/worthiness";
+    }
+}
