+++ src/Services/Asgard/Services/ICurrentHeroService.cs
@@ -1,0 +1,7 @@
+namespace Asgard.Services
+{
+    public interface ICurrentHeroService
+    {
+        bool WieldAsgard();
+    }
+}
