+++ src/Services/Heimdall/Options/JwtOptions.cs
@@ -1,0 +1,9 @@
+namespace Heimdall.Options
+{
+    public class JwtOptions
+    {
+        public const string Key = "Jwt";
+
+        public string Secret { get; set; }
+    }
+}
