--- tests/Mjolnir.Tests/Acceptance/Acceptance.csproj
@@ -1,25 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="3.1.22" />
-    <PackageReference Include="FluentAssertions" Version="5.10.3" />
-    <PackageReference Include="MartinCostello.Logging.XUnit" Version="0.2.0" />
-    <PackageReference Include="SpecFlow.Plus.LivingDocPlugin" Version="3.9.*" />
-    <PackageReference Include="SpecFlow.xUnit" Version="3.9.52" />
-    <PackageReference Include="xunit" Version="2.4.1" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.1">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\src\Mjolnir\Mjolnir.csproj" />
-  </ItemGroup>
-
-</Project>
