--- tests/Mjolnir.Tests/Integration/UnitTest1.cs
@@ -1,14 +1,0 @@
-using System;
-using Xunit;
-
-namespace Integration
-{
-    public class UnitTest1
-    {
-        [Fact]
-        public void Test1()
-        {
-
-        }
-    }
-}
